+++ typescript/infra/src/config/funding/alert-query-templates.ts
